requirements.txt
asgiref==3.9.1
Django==5.2.6
django-bootstrap5==24.2
django-debug-toolbar==4.2.0
pillow==10.4.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
